requirements.txt
alabaster==0.7.12
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.3
Django==3.1
django-datatable==0.3.1
django-templated-mail==1.1.1
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-datatables==0.5.2
djoser==2.1.0
docutils==0.16
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.4
psycopg2==2.8.6
Pygments==2.6.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0